serverとclientの共通の関数を切り出した
