+++ echo_server/server.c
@@ -8,29 +8,10 @@
 #include <stdlib.h>
 #include <errno.h>
 #define BUFFER_SIZE 1024
+#include "utils.h"
 volatile int client_num = 0;
 volatile sig_atomic_t server_running = 1;
 int socket_fd;
-
-int send_all(int socket_fd, char *buffer, int length)
-{
-    int total_sent = 0;
-    int bytes_remain = length;
-    ssize_t bytes_sent;
-
-    while (total_sent < length)
-    {
-        bytes_sent = send(socket_fd, buffer + total_sent, bytes_remain, 0);
-        if (bytes_sent == -1)
-        {
-            break;
-        }
-        total_sent += bytes_sent;
-        bytes_remain -= bytes_sent;
-    }
-
-    return total_sent;
-}
 
 void handle_sigint(int sig)
 {
